Power_of_2.cpp
This C++ program, authored by ADARSH, efficiently checks whether a user-inputted number is a power of 2 or not using bitwise manipulation. It employs the bitwise AND operation with n and n - 1 to clear the rightmost set bit. If the result is 0 and n is not zero, the program determines that the input number has only one bit set, indicating it is a power of 2. The program prompts the user for a number, performs this check, and then outputs either "It's a power of 2." or "It's not a power of 2." based on the evaluation, providing an immediate and clear response about the nature of the input number.